requirements.txt
pyodbc>=5.1,<6.0
python-dotenv>=1.0,<2.0

